Improved modularity and added temporary directories for data structuring
